Set ContentLength explicitly in test for go1.7 compatibility
Something changed between go1.7 and go1.9. In the earlier version,
net/http's client would refuse to send a request at all if the
ContentLength of the request did not match the number of bytes it got
out of the writer. This broke a test that wanted to send a truncated
request body.