Return Close() errors instead of logging
It's best not to call the standard library's logger directly without
any configuration, so this changes clients to return errors
encountered when calling resp.Body.Close() instead of logging them.

If the client encounters a different error, any error from Close()
will be ignored. If Close() returns an error, we'll call it a second
time due to the defer, but this should be harmless.