Make the python client code ascii-encode server urls.
Without this, if passed a unicode string object as the address for the
Twirp server, the client will break on python2 if a proto message
definition has a field with id 16 or greater, if it contains varints
(int32, etc) with value over 127, or simply has binary byte fields with
non-ascii characters.

The root of the problem is that Python2's unicode types are infectious.
If the url passed is a unicode() type and not a str() type, urllib will
silently make the entire set of headers unicode when the Host header is
appended.   When the protobuf encoded body is appended later, it will
attempt to encode it as unicode and may fail.

The error message you'll get will look something like this:

```
  File "/usr/lib/python2.7/httplib.py", line 1038, in endheaders
    self._send_output(message_body)
  File "/usr/lib/python2.7/httplib.py", line 880, in _send_output
    msg += message_body
UnicodeDecodeError: 'ascii' codec can't decode byte 0x92 in position 170: ordinal not in range(128)
```

This solves the problem by forcibly encoding the URL of the server in
ASCII, which matches what is allowed to be in DNS names and URLs.