Support proto3 field presence
Protobuf recently added support for [optional fields][].
These are opt-in on a per-code-generator basis.
Without this flag, you get an error:

    example.proto: is a proto3 file that contains optional fields, but code generator protoc-gen-twirp hasn't been updated to support optional fields in proto3. Please ask the owner of this code generator to support proto3 optional.

As far as I can tell, there should not be anything special we need to do
to support optional fields in Twirp as we only reference message types
and don't do anything with their fields.

[optional fields]: https://github.com/protocolbuffers/protobuf/blob/v3.12.0/docs/field_presence.md#application-note-field-presence